🔨 building now works for index project
