+++ blog-index/package.json
@@ -1,4 +1,5 @@
 {
     "name": "@try/blog-index",
-    "version": "0.0.0"
+    "version": "0.0.0",
+    "type": "module"
 }
