+++ brand-buttons/package.json
@@ -1,4 +1,5 @@
 {
     "name": "@try/brand-buttons",
-    "version": "0.0.0"
+    "version": "0.0.0",
+    "type": "module"
 }
