+++ content-cards/package.json
@@ -1,4 +1,5 @@
 {
     "name": "@try/content-cards",
-    "version": "0.0.0"
+    "version": "0.0.0",
+    "type": "module"
 }
