+++ girlscript-portfolio-page/package.json
@@ -1,4 +1,5 @@
 {
     "name": "@try/girlscript-portfolio-page",
-    "version": "0.0.0"
+    "version": "0.0.0",
+    "type": "module"
 }
