+++ index/package.json
@@ -1,4 +1,9 @@
 {
     "name": "@try/index",
-    "version": "0.0.0"
+    "version": "0.0.0",
+    "type": "module",
+    "scripts": {
+        "build": "vite --config vite.config.js build src",
+        "dev": "vite --config vite.config.js src"
+    }
 }
