+++ index/src/index.html
@@ -7,8 +7,7 @@
     <title>Try out — Vasanth Srivatsa</title>
 
     <script src="https://cdn.tailwindcss.com"></script>
-    <script type="module" src="/index/index.js"></script>
-    <script type="module" src="/index/tailwind.config.js"></script>
+    <script type="module" src="./index.js"></script>
 
     <!-- importing vyaktitva -->
     <link rel="preconnect" href="https://unpkg.com">
@@ -17,7 +16,7 @@
     <link rel="stylesheet" href="https://unpkg.com/vyaktitva/dist/vyaktitva/vyaktitva.css">
     <script type="module" async src="https://unpkg.com/vyaktitva/dist/vyaktitva/vyaktitva.esm.js"></script>
 
-    <link rel="stylesheet" href="/index/index.css">
+    <link rel="stylesheet" href="./index.css">
 </head>
 <body class="flex flex-col font-body h-screen select-none bg-light-100 dark:bg-dark-900 dark:text-white">
     <!-- header -->
