+++ package.json
@@ -17,9 +17,13 @@
         "url": "https://github.com/vsnthdev/try.git"
     },
     "scripts": {
-        "postinstall": "lerna bootstrap"
+        "postinstall": "lerna bootstrap",
+        "build": "lerna run build",
+        "clean": "lerna run clean"
     },
     "devDependencies": {
+        "axios": "^0.24.0",
+        "deepmerge": "^4.2.2",
         "lerna": "^4.0.0",
         "vite": "^2.7.10"
     }
