+++ profile-card-sharp-shadow/package.json
@@ -1,4 +1,5 @@
 {
     "name": "@try/profile-card-sharp-shadow",
-    "version": "0.0.0"
+    "version": "0.0.0",
+    "type": "module"
 }
