+++ vite.config.js
@@ -1,0 +1,19 @@
+/*
+ *  Global Vite.js configuration used across all sub-projects.
+ *  Created On 01 January 2022
+ */
+
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+    clearScreen: false,
+    publicDir: '../public',
+    build: {
+        emptyOutDir: true
+    },
+    server: {
+        fs: {
+            strict: false
+        }
+    }
+})
